can be compiled with g++

+++ kstring.c
@@ -12,7 +12,7 @@ int ksprintf(kstring_t *s, const char *fmt, ...)
 	va_start(ap, fmt);
 	l = vsnprintf(s->s + s->l, s->m - s->l, fmt, ap); // This line does not work with glibc 2.0. See `man snprintf'.
 	va_end(ap);
-	if (l + 1 > s->m - s->l) {
+	if ((size_t)l + 1 > s->m - s->l) {
 		s->m = s->l + l + 2;
 		kroundup32(s->m);
 		s->s = (char*)realloc(s->s, s->m);
